services/api/pyproject.toml
[project]
name = "showcase-api"
version = "0.1.0"
description = "Backend API для витрины товаров"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "telegram-init-data>=0.2.0",
    "python-multipart>=0.0.12",
    "python-json-logger>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
